pyproject.toml
[tool.poetry]
name = "kumiko-gws-scraper"
version = "0.1.0"
description = ""
authors = ["No767 <73260931+No767@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
aiodns = "^3.0.0"
Brotli = "^1.0.9"
orjson = "^3.7.12"
cchardet = "^2.1.7"
uvloop = "^0.16.0"
lxml = "^4.9.1"
beautifulsoup4 = "^4.11.1"
SQLAlchemy = "^1.4.40"
asyncpg = "^0.26.0"
python-dotenv = "^0.20.0"
numpy = "^1.23.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
